requirements.txt
flask
uwsgi
Flask-Session
requests
beautifulsoup4
python-dotenv
cachelib
PyJWT
lxml
cryptography